+++ src/effects/feats/sacredWeapon.js
@@ -19,7 +19,7 @@ export async function sacredWeaponEffect(document) {
   //   },
   // );
 
-  if (CONFIG.DDBI.EFFECT_CONFIG.FEATS.installedModules.atlInstalled) {
+  if (CONFIG.DDBI.EFFECT_CONFIG.MODULES.installedModules.atlInstalled) {
     effect.changes.push(generateATLChange("ATL.dimLight", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '5'));
     effect.changes.push(generateATLChange("ATL.lightColor", CONST.ACTIVE_EFFECT_MODES.OVERRIDE, '#ffffff'));
     effect.changes.push(generateATLChange("ATL.lightAlpha", CONST.ACTIVE_EFFECT_MODES.OVERRIDE, '0.25'));