+++ src/effects/specialSpells.js
@@ -179,8 +179,8 @@ export async function spellEffectAdjustment(document) {
     logger.warn("Sorry, you're missing some required modules for spell effects. Please install them and try again.", deps);
     return document;
   }
-  if (!CONFIG.DDBI.EFFECT_CONFIG.SPELLS.configured) {
-    CONFIG.DDBI.EFFECT_CONFIG.SPELLS.configured = configureDependencies();
+  if (!CONFIG.DDBI.EFFECT_CONFIG.MODULES.configured) {
+    CONFIG.DDBI.EFFECT_CONFIG.MODULES.configured = configureDependencies();
   }
 
   const name = document.flags.ddbimporter.originalName ?? document.name;